+++ .editorconfig
@@ -1,0 +1,6 @@
+# .editorconfig
+root = true
+
+[*]
+end_of_line = lf
+insert_final_newline = true
