+++ examples/recaptchav2_request.py
@@ -1,13 +1,15 @@
 import requests
 from bs4 import BeautifulSoup
-from capmonster_python import RecaptchaV2Task
+
+from src.capmonster_python import RecaptchaV2Task
 
 
 class RecaptchaRequest:
     def __init__(self, _client_key: str):
         self.captcha = RecaptchaV2Task(_client_key)
         self.s = requests.Session()
-        self.s.headers = {"User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64; rv:91.0) Gecko/20100101 Firefox/91.0"}
+        self.s.headers = {
+            "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64; rv:91.0) Gecko/20100101 Firefox/91.0"}
         self.website_url = "https://google.com/recaptcha/api2/demo"
         self.expected = "Verification Success... Hooray!"
 
@@ -33,6 +35,7 @@ class RecaptchaRequest:
 
 if __name__ == "__main__":
     from os import environ
+
     client_key = environ["API_KEY"]
     example_request = RecaptchaRequest(client_key)
     assert example_request.expected in example_request.submit_form()
