+++ examples/recaptchav2_selenium.py
@@ -1,10 +1,12 @@
+from time import sleep
+
 from selenium import webdriver
 from selenium.webdriver.common.by import By
 from selenium.webdriver.firefox.options import Options
 from selenium.webdriver.firefox.service import Service
 from webdriver_manager.firefox import GeckoDriverManager
-from capmonster_python import RecaptchaV2Task
-from time import sleep
+
+from src.capmonster_python import RecaptchaV2Task
 
 
 class RecaptchaV2Selenium:
@@ -41,6 +43,7 @@ class RecaptchaV2Selenium:
 
 if __name__ == "__main__":
     from os import environ
+
     client_key = environ["API_KEY"]
     headless = environ["HEADLESS"]
     environ["WDM_LOG_LEVEL"] = "0"
