--- requirements.txt
@@ -1,2 +1,0 @@
-capmonster_python~=3.2.0
-requests~=2.32.3
