+++ src/capmonster_python/tasks/binance.py
@@ -2,7 +2,7 @@ from typing import Optional, Any
 
 from pydantic import Field
 
-from capmonster_python.tasks.task import TaskPayload, UserAgentPayload, ProxyPayload
+from .task import TaskPayload, UserAgentPayload, ProxyPayload
 
 
 class BinanceTask(TaskPayload, UserAgentPayload):
