+++ src/capmonster_python/tasks/imperva.py
@@ -2,7 +2,7 @@ from typing import Optional, Any
 
 from pydantic import Field, BaseModel
 
-from capmonster_python.tasks.task import TaskPayload, UserAgentPayload
+from .task import TaskPayload, UserAgentPayload
 
 
 class ImpervaTaskMetadata(BaseModel):
