+++ src/capmonster_python/tasks/recaptcha_v3.py
@@ -2,7 +2,7 @@ from typing import Any, Optional
 
 from pydantic import Field
 
-from capmonster_python.tasks.task import TaskPayload
+from .task import TaskPayload
 
 
 class RecaptchaV3Task(TaskPayload):
