+++ src/capmonster_python/tasks/tendi.py
@@ -2,7 +2,7 @@ from typing import Optional, Any
 
 from pydantic import Field
 
-from capmonster_python.tasks.task import TaskPayload, UserAgentPayload, ProxyPayload
+from .task import TaskPayload, UserAgentPayload, ProxyPayload
 
 
 class TenDITask(TaskPayload, UserAgentPayload):
